requirements.txt
tk
Pillow



